refactor(backend): remove extra dot in base dto filename
